% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_atlas.R
\name{connect_atlas_db}
\alias{connect_atlas_db}
\title{create a database connection to FROIAtlas.sqlite db}
\usage{
connect_atlas_db(dbpath = NULL)
}
\description{
create a database connection to FROIAtlas.sqlite db
}
