% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_atlas.R
\name{get_roi_foci}
\alias{get_roi_foci}
\title{get table of information for supplied region of interest}
\usage{
get_roi_foci(conn, froi, hemi = NULL)
}
\description{
get table of information for supplied region of interest
}
