% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_atlas.R
\name{outliers}
\alias{outliers}
\title{get outlying coordinates based on multivariate test from \code{mvoutlier::sign1}}
\usage{
outliers(coords, qcrit = 0.999, plot = TRUE)
}
\description{
get outlying coordinates based on multivariate test from \code{mvoutlier::sign1}
}
